Added filter methods for amount and credit type in transactions test

+++ library/src/main/java/com/thoughtworks/DebitTransaction.java
@@ -3,11 +3,13 @@ package com.thoughtworks;
 import java.util.Date;
 
 public class DebitTransaction extends Transaction{
-    protected DebitTransaction(Date date, float amount, AccountNumber To) {
+
+    protected DebitTransaction(Date date, double amount, AccountNumber To) {
         super(date,amount,To);
     }
 
-    public DebitTransaction(float amount, AccountNumber To) {
+    public DebitTransaction(double amount, AccountNumber To) {
         this(new Date(),amount,To);
     }
+
 }
